@@ -1,51 +1,42 @@
 {
   "name": "webrtc-video-room",
-  "version": "1.0.0",
-  "description": "WebRTC Video Room is a Single-Page App in React using React Router that provides peer-to-peer video and audio communication in a web browser with no plugins or extensions required.",
+  "version": "2.0.0",
+  "description": "WebRTC Video Room is a Single-Page App in React using React Router that provides peer-to-peer video and audio communication in a web browser with minimum dependencies required.",
   "main": "index.js",
   "repository": {
     "type": "git",
     "url": "https://github.com/dondido/webrtc-video-room.git"
   },
   "scripts": {
-    "start": "node server.js",
-    "prod": "NODE_ENV=production node server.js",
-    "postinstall": "webpack -p --config ./webpack.config.js --progress"
+    "postinstall": "webpack --mode production"
   },
   "author": "Dian Dimitrov",
+  "keywords": [
+    "WebRTC Video Chat",
+    "RTCPeerConnection",
+    "React"
+  ],
   "license": "MIT",
-  "dependencies": {
-    "babel-loader": "^6.2.5",
-    "babel-plugin-transform-class-properties": "^6.11.5",
-    "babel-plugin-transform-runtime": "^6.12.0",
-    "babel-polyfill": "^6.3.14",
-    "babel-preset-es2015": "^6.3.13",
-    "babel-preset-react": "^6.3.13",
-    "babel-preset-stage-0": "^6.3.13",
-    "babel-register": "^6.4.3",
-    "babel-runtime": "^6.11.6",
-    "compression": "^1.6.2",
-    "css-loader": "^0.25.0",
-    "express": "^4.13.4",
-    "express-session": "^1.14.1",
-    "react": "^15.3.1",
-    "react-dom": "^15.3.1",
-    "react-redux": "^4.4.5",
-    "react-router": "^2.7.0",
-    "redux": "^3.6.0",
-    "remarkable": "^1.6.2",
-    "serve-favicon": "^2.3.0",
-    "socket.io": "^1.4.8",
-    "style-loader": "^0.13.1",
-    "webpack": "^1.13.2"
-  },
   "devDependencies": {
-    "babel-eslint": "^7.0.0",
-    "eslint": "^3.7.0",
-    "eslint-config-airbnb": "^12.0.0",
-    "eslint-loader": "^1.5.0",
-    "eslint-plugin-import": "^2.0.0",
-    "eslint-plugin-jsx-a11y": "^2.2.2",
-    "eslint-plugin-react": "^6.3.0"
+    "babel-core": "^6.26.3",
+    "babel-loader": "^7.1.4",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-react": "^6.24.1",
+    "css-loader": "^0.28.11",
+    "style-loader": "^0.21.0",
+    "webpack": "^4.10.2",
+    "webpack-cli": "^3.0.2"
+  },
+  "dependencies": {
+    "compression": "^1.7.2",
+    "express": "^4.16.3",
+    "react": "^16.4.0",
+    "react-dom": "^16.4.0",
+    "react-redux": "^5.0.7",
+    "react-router-dom": "^4.3.1",
+    "redux": "^4.0.0",
+    "remarkable": "^1.7.1",
+    "serve-favicon": "^2.5.0",
+    "socket.io": "^2.1.1"
   }
 }
