@@ -1,25 +1,28 @@
-import React from 'react'
-import { Link } from 'react-router'
+import React from 'react';
+import { PropTypes } from 'prop-types';
+import { connect } from 'react-redux';
+import { Link } from 'react-router-dom';
 
 const Home = props =>
   <div className="home">
     <div>
       <h1 itemProp="headline">Webrtc Video Room</h1>
       <p>Please enter a room name.</p>
-      <form onSubmit={props.joinRoom}>
-        <input type="text" name="room" value={props.roomId} onChange={props.handleChange} pattern="^\w+$" maxLength="10" required autoFocus title="Room name should only contain letters or numbers."/>
-        <button className="primary-button" type="submit">Join</button>
-        <button className="primary-button" onClick={props.setRoom}>Random</button>
-      </form>
-      {props.rooms.length !== 0 && <div>Recently used rooms:</div>}
-      {[...props.rooms].map(room => <Link key={room} className="recent-room" to={'/r/' + room}>{room}</Link>)}
+      <input type="text" name="room" value={ props.roomId } onChange={props.handleChange} pattern="^\w+$" maxLength="10" required autoFocus title="Room name should only contain letters or numbers."/>
+      <Link className="primary-button" to={ '/r/' + props.roomId }>Join</Link>
+      <Link className="primary-button" to={ '/r/' + props.defaultRoomId }>Random</Link>
+      { props.rooms.length !== 0 && <div>Recently used rooms:</div> }
+      { props.rooms.map(room => <Link key={room} className="recent-room" to={ '/r/' + room }>{ room }</Link>) }
     </div>
   </div>;
+
 Home.propTypes = {
-  handleChange: React.PropTypes.func.isRequired,
-  joinRoom: React.PropTypes.func.isRequired,
-  setRoom: React.PropTypes.func.isRequired,
-  roomId: React.PropTypes.string.isRequired,
-  rooms: React.PropTypes.object.isRequired
+  handleChange: PropTypes.func.isRequired,
+  defaultRoomId: PropTypes.string.isRequired,
+  roomId: PropTypes.string.isRequired,
+  rooms: PropTypes.array.isRequired
 };
-export default Home;
+
+const mapStateToProps = store => ({rooms: store.rooms});
+
+export default connect(mapStateToProps)(Home);
