@@ -1,34 +1,30 @@
-import React from 'react'
-import { connect } from 'react-redux'
-import Home from '../components/Home'
-import store from '../store'
-class HomePage extends React.Component {
+import React, { Component } from 'react'
+import { PropTypes } from 'prop-types';
+import Home from '../components/Home';
+
+class HomePage extends Component {
   constructor(props) {
     super(props);
+    this.defaultRoomId = String(new Date() - new Date().setHours(0, 0, 0, 0));
+    this.state = { roomId: this.defaultRoomId };
+    this.handleChange = this.handleChange.bind(this);
   }
-  state = {
-    value: new Date() - new Date().setHours(0, 0, 0, 0)
+  handleChange(e) {
+    this.setState({ roomId: e.target.value });
   }
-  static contextTypes = {
-    router: React.PropTypes.object
-  }
-  setRoom = () => this.setState({value: new Date() - new Date().setHours(0, 0, 0, 0)})
-  joinRoom = e => {
-    e.preventDefault();
-    this.context.router.push('r/' + this.state.value);
-  }
-  handleChange = e => this.setState({value: e.target.value})
   render(){
     return (
       <Home
-        roomId={this.state.value}
+        defaultRoomId={this.defaultRoomId}
+        roomId={this.state.roomId}
         handleChange={this.handleChange}
-        joinRoom={this.joinRoom}
-        setRoom={this.setRoom}
-        rooms={this.props.rooms}
-        ></Home>
+      />
     );
   }
 }
-const mapStateToProps = store => ({rooms: new Set([...store.rooms])});
-export default connect(mapStateToProps)(HomePage);
+
+HomePage.contextTypes = {
+  router: PropTypes.object
+};
+
+export default HomePage;
