@@ -1,7 +1,6 @@
-const initialState = new Set();
-const updateRooms = (state = initialState, action) => {
+const updateRooms = (state = [], action) => {
   if (action.type === 'ADD_ROOM') {
-    return new Set([...state, action.room]);
+    return [...new Set([...state, action.room])];
   }
   return state;
 };
