@@ -1,35 +1,22 @@
-// webpack should be in the node_modules directory, install if not.
-// https://medium.com/@tkssharma/eslint-in-react-babel-webpack-9cb1c4e86f4e#.hlophrwed
-var webpack = require("webpack");
+const path = require("path");
+
 module.exports = {
-  devtool: /*'eval',*/'cheap-module-source-map',
-  entry: './src/index.js',
+  entry: "./src/index.js",
   output: {
-    path: './public',
-    filename: 'bundle.js'
+    path: path.join(__dirname, "/dist"),
+    filename: "./bundle.js"
   },
   module: {
-    loaders: [
+    rules: [
       {
         test: /\.js$/,
         exclude: /node_modules/,
-        loaders: ['babel-loader', 'eslint-loader']
+        use: ["babel-loader"]
       },
       {
-        test: /\.css/,
-        loaders: ['style', 'css'],
-        include: __dirname + '/src'
+        test: /\.css$/,
+        use: ["style-loader", "css-loader"]
       }
     ]
-  },
-  plugins:[
-    new webpack.DefinePlugin({
-      'process.env': {
-        'NODE_ENV': JSON.stringify('production')
-      }
-    })
-  ],
-  resolve: {
-    extensions: ['', '.js', '.json']
   }
 };
